refactor: move area domain related class to exclusive package
