+++ src/main/java/com/fhsbatista/journal/data/area/AreaRepository.java
@@ -1,6 +1,6 @@
 package com.fhsbatista.journal.data.area;
 
-import com.fhsbatista.journal.domain.Area;
+import com.fhsbatista.journal.domain.area.Area;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AreaRepository extends JpaRepository<Area, Long> {
