+++ src/main/java/com/fhsbatista/journal/presentation/area/dto/AreaRegisterDTO.java
@@ -1,6 +1,6 @@
 package com.fhsbatista.journal.presentation.area.dto;
 
-import com.fhsbatista.journal.domain.Area;
+import com.fhsbatista.journal.domain.area.Area;
 import jakarta.validation.constraints.NotBlank;
 
 public record AreaRegisterDTO(
