+++ src/main/java/com/fhsbatista/journal/presentation/habit/HabitCreateDTO.java
@@ -1,6 +1,6 @@
 package com.fhsbatista.journal.presentation.habit;
 
-import com.fhsbatista.journal.domain.Area;
+import com.fhsbatista.journal.domain.area.Area;
 import com.fhsbatista.journal.domain.Habit;
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.DecimalMin;
