+++ src/test/java/com/fhsbatista/journal/domain/HabitTest.java
@@ -1,5 +1,6 @@
 package com.fhsbatista.journal.domain;
 
+import com.fhsbatista.journal.domain.area.Area;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDate;
